+++ CHANGELOG.rst
@@ -21,3 +21,4 @@ Changelog
 
 * Improved code quality
 * Add Entity.custom_properties
+* Simplify node access (energy_system.nodes[label])